+++ src/main/java/com/example/repository/AddressRepo.java
@@ -3,6 +3,6 @@ package com.example.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.entity.Address;
 
-public interface AddressRepo extends JpaRepository<Address, Integer> {
+public interface AddressRepo extends JpaRepository<Address, Long> {
 
 }
